+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## x.x.x.x
 - Added setting to turn on/off ability to change volume with the scroll wheel anywhere (thanks @Tester798!)
 - Added setting to turn on/off ability to change volume with the scroll wheel when hovering over the EarTrumpet icon (thanks @Tester798!)
+- Added new community settings area
+- Added new community setting to turn on/off use of a logarithmic volume scale (thanks @yonatan-mitmit!)
 - Added legacy shortcuts to the context menu pointing to [App volume and device preferences] / [Volume mixer]
 - Added ability to use the Windows key in shortcuts (thanks @iamevn!)
 - Added linguistic display name sorting for audio devices (thanks @Tester798!)